requirements.txt
# Core dependencies
streamlit>=1.29.0
pandas>=2.2.0
numpy>=1.26.0

# Database
mysql-connector-python>=8.2.0
SQLAlchemy>=2.0.23

# AI/ML
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.6.1
scikit-learn>=1.3.2
scipy>=1.10.0
joblib>=1.2.0

# API Integration
requests>=2.31.0
googlemaps>=4.10.0

# Utilities
python-dotenv>=1.0.0
folium>=0.15.1
streamlit-folium>=0.15.1

# Testing and plotting (development)
pytest>=7.0.0
matplotlib>=3.6.0

# Optional (requires compilation / binaries): uncomment to enable faster GBMs
# lightgbm>=4.4.0
# xgboost>=1.7.0
